+++ app/src/components/Post.js
@@ -1,9 +1,10 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import moment from 'moment'
-import { withRouter, Link } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import { Label, Button } from 'semantic-ui-react'
 import { createSkeletonElement } from '@trainline/react-skeletor'
 import LikeBtn from './../pages/InsightsNew/LikeBtn'
+import PostVisualBacktest from './PostVisualBacktest'
 import './Post.css'
 
 const A = createSkeletonElement('a', 'pending-home')
@@ -44,15 +45,11 @@ const Status = ({ status = STATES.draft, moderationComment }) => {
 
 const Author = ({ id, username }) => (
   <div className='event-post-author'>
-    {id && (
-      <Fragment>
-        by&nbsp; <Link to={`/insights/users/${id}`}>{username}</Link>
-      </Fragment>
-    )}
+    by&nbsp; <Link to={`/insights/users/${id}`}>{username}</Link>
   </div>
 )
 
-const Post = ({
+export const Post = ({
   index = 1,
   id,
   title,
@@ -67,10 +64,10 @@ const Post = ({
   unvotePost,
   deletePost,
   publishPost,
-  history,
   moderationComment = null,
   state = STATES.approved,
   readyState = STATES.draft,
+  discourseTopicUrl = '',
   gotoInsight,
   showStatus = false
 }) => {
@@ -89,10 +86,10 @@ const Post = ({
         </A>
         <br />
         <Span>{moment(createdAt).format('MMM DD, YYYY')}</Span>
-        {user && tags.length > 0 && <Author {...user} />}
+        {user && user.id && <Author {...user} />}
         {user && (
           <Div className='event-post-info'>
-            {tags.length > 0 ? (
+            {tags.length > 0 && (
               <div className='post-tags'>
                 {tags.map((tag, index) => (
                   <Link
@@ -104,8 +101,6 @@ const Post = ({
                   </Link>
                 ))}
               </div>
-            ) : (
-              <Author {...user} />
             )}
             <LikeBtn
               onLike={() => {
@@ -117,21 +112,27 @@ const Post = ({
               }}
               balance={balance}
               liked={!!votedAt}
-              votes={votes.totalSanVotes || 0}
+              votes={votes}
             />
           </Div>
         )}
+        {discourseTopicUrl && (
+          <a className='discussion-btn' href={discourseTopicUrl}>
+            go to discussion
+          </a>
+        )}
+        <br />
+        {createdAt &&
+          tags.length > 0 && (
+            <PostVisualBacktest from={createdAt} ticker={tags[0].name} />
+          )}
         <Div className='event-post-controls'>
           {showStatus && (
             <Status moderationComment={moderationComment} status={readyState} />
           )}
-          <div
-            style={{
-              display: 'flex'
-            }}
-          >
-            {showStatus &&
-              readyState === 'draft' && (
+          {showStatus &&
+            readyState === 'draft' && (
+              <div style={{ display: 'flex' }}>
                 <Button
                   size='mini'
                   onClick={() => deletePost(id)}
@@ -143,9 +144,6 @@ const Post = ({
                 >
                   Delete this insight
                 </Button>
-              )}
-            {showStatus &&
-              readyState === 'draft' && (
                 <Button
                   size='mini'
                   color='orange'
@@ -156,14 +154,12 @@ const Post = ({
                 >
                   Publish your insight
                 </Button>
-              )}
-          </div>
+              </div>
+            )}
         </Div>
       </div>
     </div>
   )
 }
 
-export const UnwrappedPost = Post // for tests
-
-export default withRouter(Post)
+export default Post